bid-analysis-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1

# parsing
pypdf==5.3.0
python-docx==1.1.2
pymupdf==1.24.13
pytesseract==0.3.13
Pillow==11.1.0

# similarity
scikit-learn==1.6.1

